[dev-client] Fix iOS unit tests
